[Python] Add UnboundedSource ValidatesRunner test
Add test_unbounded_source_read to the base FnApiRunnerTest so the
in-process FnApiRunner variants and the portable ValidatesRunner suites
exercise the UnboundedSource SDF wrapper end to end: read a
self-terminating source, assert the elements and that the EOF
MAX_TIMESTAMP watermark lets a downstream FixedWindows + GroupByKey
fire.

Move UnboundedCountingSource from unbounded_source_test to
unbounded_source as an internal helper so the runner tests and
iobase_test share it without importing across test modules.

The multi-worker and bundle-repeat FnApiRunner variants skip the test as
single worker only, matching the sibling SDF checkpointing and watermark
tests. SparkRunnerTest skips it because portable Spark does not execute
SDFs (#19468).

Bump the Dataflow, Flink and Spark Python ValidatesRunner trigger files
so the PostCommit suites run on this PR.
